Move db methods to runningtracker.db package
