+++ runningtracker/db/db_interface.py
@@ -2,7 +2,7 @@ from sqlite3 import connect, Connection
 from pathlib import Path
 SCHEMA_VERSION = 'v1'
 SCHEMA_LOCATION = str(
-    Path('.').parent.joinpath('database', f"schema_{SCHEMA_VERSION}.sql")
+    Path('.').parent.parent.joinpath('database', f"schema_{SCHEMA_VERSION}.sql")
 )
 DATABASE_NAME = 'runningtracker.db'
